Set border-radius for buttons
This will make the buttons look a bit more like pills, which is
where the design is going to.

Answering the question about the 9999px value, it’s a shortcut to
get perfectly rounded edges without falling into the trap of getting
ellipsis instead. There are [great StackOverflow][0] answers
explaining this a bit better and Tailwind CSS [seems to follow][1]
the [same technique][2].

[0]: https://stackoverflow.com/a/29966500/590525
[1]: https://stackoverflow.com/questions/31617136/avoid-elliptical-shape-in-css-border-radius#comment98361078_31617326
[2]: https://github.com/tailwindcss/tailwindcss/blob/5066a1e97b3f06b899a4539767ca2bee4bde2da9/defaultConfig.stub.js#L484