Add a full form example
This is not an extensive example of all the possibilities a form
has, but it is quite representative of the major building blocks
we’ll be using to build most of our forms.

Alex is still to look into these and to provide an updated look,
but for now this example will serve as a testing pad for the
changes he’ll suggest soon.