Add Secondary navigation to Bridget
This commits follows Berto’s work but tries a different approach.
As usual, relying on Bootstrap to do the heavy lifting for us is
the best strategy, since it exposes to many buttons and levers we
can change and tweak to get the result wanted.

In this case, Bootstrap’s variables file has quite a few that we
should tweak in order to get the tabs we want. Besides, we should
bring the markup from Bootstrap’s examples instead of trying to
customise the markup we have on Connect (or other projects).

Co-authored-by: Berto <berto@bridge-u.com>

Thanks: @doutatsu