Add default Bootstrap buttons for reference
One day, Alex was looking for an extra button for a particular
design situation, but he didn’t know if we had them. Luckily, I was
able to point him in the right direction — but that made me realise
that Bootstrap outputs a lot by default, and some of it might be
actually useful as design inspiration.

This comit brings a section with all the default buttons Bootstrap
generates by default, as a design reference. They are not to be used
in the app unless named and declared in the previous section.