Export colour variables and display them as cards
Trying to set a direction in which to show/emulate our visual
style guide, we’re trying to get a better sense of Atomic Design.

This commit introduces a simple way to directly tap into our Sass
variables and bring them into the Storybook guide we have in order
to show developers and designers our colour building blocks.