Change relative path for variables import
This change is to resolve the path for importing our bootstrap variables
file in our main bootstrap.scss. We are trying to add bridget to the
Prepare project, and there is an issue importing this file, as the file
cannot be found by CircleCI. This makes the path work according to the
scss convention for importing scss files.