Fix Jquery version on the package.json
As we are using the Jquery ^3.4.0 versioning annotation, the library on `yarnupdate bridget` caused, to automatically install a double version of jquery into the dependant project, causing several parts of the dependent system to fail. until we fix/improve either briget or the dependent project setup. We propose to statically state the version of Jquery that we use.