Removed all the vulnerabilities
Storybook removed:
Storybook seems to be used on an obsolete project restricted to this repo. Removing it shouldn't have any effect on any of our applications.

Webpack-dev-server removed:
Webpack-dev-server was used to serve Storybook. It is no needed any longer. Bridget is now only compiling the assets.

Http-server removed:
It is not used anymore since I removed Webpack-dev-server.

Set NodeJs to v.20, but it really doesn't do much in this project.

I upgraded all the other dependencies to the latest versions. No bundling errors (only 2 CSS warnings).

At the end of this action, 0 vulnerabilities were found.