Make dependabot automatically manage stale PRs
At time of writing there are 150 PRs open in BridgeU. This repo is a
big source of false positives. This commit introduces dependabot and a
quick GH action check to make it easier to clear down automatically
and reduce the noise.

Note this is use as an npm package in Prepare so security
vulnerabilities should likely be merged still.